Bayan/requirements.txt
pygame
SpeechRecognition
vosk
groq
gTTS
python-dotenv
